+++ packages/txservice/src/types.ts
@@ -1,7 +1,9 @@
-import { Logger, RequestContext } from "@connext/nxtp-utils";
+import { delay, Logger, RequestContext } from "@connext/nxtp-utils";
 import { BigNumber, BigNumberish, providers, utils } from "ethers";
 
-import { MaxBufferLengthError, TransactionBackfilled } from "./error";
+import { MaxBufferLengthError, parseError, RpcError, ServerError, TransactionBackfilled } from "./error";
+
+export const { StaticJsonRpcProvider } = providers;
 
 export type ReadTransaction = {
   chainId: number;
@@ -19,17 +21,6 @@ export type FullTransaction = {
   gasPrice: BigNumber;
   gasLimit: BigNumber;
 } & WriteTransaction;
-
-// TODO: Cache all the provider call responses, and have one singular data structure for managing that cache.
-export type CachedGas = {
-  price: BigNumber;
-  timestamp: number;
-};
-
-export type CachedTransactionCount = {
-  value: number;
-  timestamp: number;
-};
 
 /// Events
 export type TxServiceSubmittedEvent = {
@@ -123,7 +114,7 @@ type LoggableTransactionData = {
   confirmedBlockNumber?: number;
   gasPrice: string;
   gasLimit: string;
-  error: any;
+  error?: string;
   confirmations?: number;
   state?: string;
 };
@@ -131,11 +122,7 @@ type LoggableTransactionData = {
 /**
  * @classdesc A data structure for storing invariant params and managing state related to a single transaction.
  */
-export class Transaction {
-  // TODO: Temp solution - will be replaced by batching solution.
-  // How many attempts until we consider a blocking tx as taking too long.
-  public static MAX_ATTEMPTS = 99999;
-
+export class OnchainTransaction {
   // Responses, in the order of attempts made for this tx.
   public responses: providers.TransactionResponse[] = [];
 
@@ -156,9 +143,8 @@ export class Transaction {
   // the confirmation receipt.
   public minedBlockNumber = -1;
 
-  // TODO: private setter
-  // Timestamp initially set on creation, but will be updated each time submit() is called.
-  public timestamp: number = Date.now();
+  // Timestamp initially set on creation.
+  public readonly timestamp: number = Date.now();
 
   // This error will be set in the instance of a failure.
   public error: Error | undefined;
@@ -231,9 +217,10 @@ export class Transaction {
       hashes: this.hashes.length > 0 ? this.hashes : undefined,
       // Track block numbers for mine and confirm for observability.
       minedBlockNumber: this.minedBlockNumber === -1 ? undefined : this.minedBlockNumber,
-      confirmedBlockNumber: this.receipt && (this.receipt.blockNumber > this.minedBlockNumber) ? this.receipt.blockNumber : undefined,
+      confirmedBlockNumber:
+        this.receipt && this.receipt.blockNumber > this.minedBlockNumber ? this.receipt.blockNumber : undefined,
       confirmations: this.receipt?.confirmations,
-      error: this.error,
+      error: this.error ? this.error.message : undefined,
       state: this.error ? "E" : this.didFinish ? "C" : this.didMine ? "M" : this.didSubmit ? "S" : undefined,
     };
   }
@@ -261,17 +248,21 @@ export class Transaction {
 /**
  * @classdesc A data structure for managing the lifecycle of a (continuously rotating) batch of transactions.
  */
-export class TransactionBuffer extends Array<Transaction> {
+export class TransactionBuffer extends Array<OnchainTransaction> {
   public get isFull(): boolean {
     return this.maxLength ? this.length >= this.maxLength : false;
   }
 
   public get nonces(): number[] {
-    return this.map((tx) => tx.nonce);
+    const nonces = this.map((tx) => tx.nonce);
+    if (this.lastShiftedTx) {
+      nonces.unshift(this.lastShiftedTx.nonce);
+    }
+    return nonces;
   }
 
   // We use this to record the last tx that was shifted out of the buffer.
-  private lastShiftedTx: Transaction | undefined;
+  private lastShiftedTx: OnchainTransaction | undefined;
 
   /**
    * A data structure used for management of the lifecycle of on-chain transactions.
@@ -299,7 +290,7 @@ export class TransactionBuffer extends Array<Transaction> {
    * @param tx - The transaction to add to the buffer.
    * @returns number index.
    */
-  public push(tx: Transaction): number {
+  public push(tx: OnchainTransaction): number {
     if (this.isFull) {
       throw new MaxBufferLengthError({
         maxLength: this.maxLength,
@@ -359,11 +350,15 @@ export class TransactionBuffer extends Array<Transaction> {
    * Shifts a transaction from the buffer.
    * @returns The last transaction in the buffer.
    */
-  public shift(): Transaction | undefined {
+  public shift(): OnchainTransaction | undefined {
     const tx = super.shift();
     this.lastShiftedTx = tx;
     this.log();
     return tx;
+  }
+
+  public getTxByNonce(nonce: number): OnchainTransaction | undefined {
+    return this.find((tx) => tx.nonce === nonce) ?? this.lastShiftedTx;
   }
 
   private log(message?: string, context: any = {}, error = false) {
@@ -379,6 +374,309 @@ export class TransactionBuffer extends Array<Transaction> {
       this.logger.error(msg, undefined, undefined, ctx);
     } else {
       this.logger.debug(msg, undefined, undefined, ctx);
+    }
+  }
+}
+
+/**
+ * Cache item used in ProviderCache schema property; either blocks to live (BTL) or time to live (TTL) must
+ * be specified (the former is preferred but the latter can be used as a backup when a block listener will not
+ * be actively updating the cache's block number).
+ */
+type ProviderCacheSchema<T> = {
+  [K in keyof T]:
+    | {
+        btl: number;
+        ttl?: number;
+      }
+    | {
+        btl?: number;
+        ttl: number;
+      };
+};
+
+type ProviderCacheData<T> = {
+  [K in keyof T]?: {
+    value: T[K];
+    timestamp: number;
+    blockNumber: number;
+  };
+};
+
+/**
+ * @classdesc A data structure for managing time-sensitive (expiring) cached information from chain
+ * that expires after a set number of blocks or amount of time.
+ */
+export class ProviderCache<T> {
+  private _blockNumber = -1;
+  public get blockNumber(): number {
+    return this._blockNumber;
+  }
+
+  private _data: ProviderCacheData<T> = {};
+  public get data(): Partial<T> {
+    const data: Partial<T> = {};
+    for (const k of Object.keys(this._data)) {
+      const key = k as keyof T;
+      data[key] = this.getItem(key)?.value;
+    }
+    return data;
+  }
+
+  /**
+   * @param schema - A schema for the cache that determines whether each item expires after a set period of
+   * time (ttl, ms) or a set number of blocks (btl, number).
+   */
+  constructor(private readonly logger: Logger, private readonly schema: ProviderCacheSchema<T>) {}
+
+  /**
+   * Update the cache block number, and optionally the data.
+   * @param blockNumber - Current block number.
+   * @param data - Optional data to update the cache with.
+   */
+  public update(blockNumber: number, data: Partial<T> = {}) {
+    if (blockNumber < this._blockNumber) {
+      this.logger.debug("Block number went backwards. Did a reorg occur?", undefined, undefined, {
+        newBlockNumber: blockNumber,
+        previousBlockNumber: this._blockNumber,
+      });
+    }
+    this._blockNumber = blockNumber;
+    this.set(data);
+  }
+
+  /**
+   * Set a value in the cache.
+   * @param data - The data to set.
+   */
+  public set(data: Partial<T>) {
+    Object.keys(data).forEach((k) => {
+      const key = k as keyof T;
+      const value = data[key] as T[keyof T];
+      this._data[key] = {
+        value,
+        timestamp: Date.now(),
+        blockNumber: this.blockNumber,
+      };
+    });
+  }
+
+  /**
+   * Helper for retrieving item from data depending on whether it's expired.
+   * @param key - a key of the cache data schema.
+   * @returns
+   */
+  private getItem(key: keyof T): ProviderCacheData<T>[keyof T] | undefined {
+    const { ttl, btl } = this.schema[key];
+    const item = this._data[key];
+    if (!item) {
+      return undefined;
+    }
+    // In these blocks, we'll also erase the item from the cache data if it's expired.
+    if (ttl !== undefined && item.timestamp + ttl < Date.now()) {
+      this._data[key] = undefined;
+      return undefined;
+    }
+    if (btl !== undefined && item.blockNumber + btl < this.blockNumber) {
+      this._data[key] = undefined;
+      return undefined;
+    }
+    return item;
+  }
+}
+
+/**
+ * @classdesc An extension of StaticJsonRpcProvider that manages a providers chain synchronization status
+ * and intercepts all RPC send() calls to ensure that the provider is in sync.
+ */
+export class SyncProvider extends StaticJsonRpcProvider {
+  private readonly connectionInfo: utils.ConnectionInfo;
+  public readonly url: string;
+
+  public synced = true;
+  public lag = 0;
+  public priority = 0;
+
+  private static readonly N_SAMPLES = 100;
+  // Denominator is the target reliability sample size.
+  private static readonly RELIABILITY_STEP = 1 / SyncProvider.N_SAMPLES;
+  // A metric used for measuring reliability, based on the number of successful calls / last N calls made.
+  public reliability = 0.0;
+
+  // Used for tracking how many calls we've made in the last second.
+  private cpsTimestamps: number[] = [];
+  public get cps(): number {
+    // Average CPS over the last 10 seconds.
+    const now = Date.now();
+    this.cpsTimestamps = this.cpsTimestamps.filter((ts) => now - ts < 10_000);
+    return this.cpsTimestamps.length / 10;
+  }
+  private execTimes: number[] = [];
+  public get avgExecTime(): number {
+    // Average execution time over the last N samples.
+    this.execTimes = this.execTimes.slice(-SyncProvider.N_SAMPLES);
+    return this.execTimes.reduce((a, b) => a + b, 0) / this.execTimes.length;
+  }
+
+  // This variable is used to track the last block number this provider synced to, and is kept separately from the
+  // inherited `blockNumber` property (which is a getter that uses an update method).
+  private _syncedBlockNumber = -1;
+  public get syncedBlockNumber(): number {
+    return this._syncedBlockNumber;
+  }
+
+  constructor(
+    _connectionInfo: utils.ConnectionInfo | string,
+    public readonly chainId: number,
+    private readonly stallTimeout = 10_000,
+    private readonly debugLogging = true,
+  ) {
+    super(_connectionInfo, chainId);
+    this.connectionInfo = typeof _connectionInfo === "string" ? { url: _connectionInfo } : _connectionInfo;
+    this.url = this.connectionInfo.url.replace("https://", "").split(".com")[0];
+  }
+
+  /**
+   * Synchronizes the provider with chain by checking the current block number and updating the syncedBlockNumber
+   * property.
+   */
+  public sync(): Promise<void> {
+    this.removeAllListeners();
+
+    return new Promise<void>((resolve) => {
+      this.once("block", async (blockNumber: number) => {
+        this.debugLog("SYNCING_BLOCK_EVENT", blockNumber, this.syncedBlockNumber);
+        this._syncedBlockNumber = blockNumber;
+        resolve();
+      });
+    });
+  }
+
+  public async addBlockListener(onBlock: (blockNumber: number) => void) {
+    this.on("block", async (blockNumber: number) => {
+      this.debugLog("BLOCK_EVENT", blockNumber);
+      onBlock(blockNumber);
+    });
+  }
+
+  /**
+   * Overridden RPC send method. If the provider is currently out of sync, this method will
+   * now throw an RpcError indicating such. This way, we ensure an out of sync provider is never
+   * consulted (except when checking the block number, which is used for syncing).
+   *
+   * @param method - RPC method name.
+   * @param params - RPC method params.
+   * @returns any - RPC response.
+   * @throws RpcError - If the provider is currently out of sync.
+   */
+  public async send(method: string, params: Array<any>): Promise<any> {
+    // provider.ready returns a Promise which will stall until the network has been established, ignoring
+    // errors due to the target node not being active yet. This will ensure we wait until the node is up
+    // and running smoothly.
+    if ((!this.synced && method !== "eth_blockNumber") || !(await this.ready)) {
+      throw new RpcError(RpcError.reasons.OutOfSync, {
+        provider: this.url,
+        chainId: this.chainId,
+        blockNumber: this.syncedBlockNumber,
+        synced: this.synced,
+      });
+    }
+
+    // TODO: Make # of retries configurable?
+    const errors = [];
+    let sendTimestamp = -1;
+    for (let i = 1; i <= 5; i++) {
+      try {
+        sendTimestamp = Date.now();
+        this.cpsTimestamps.push(sendTimestamp);
+        return await Promise.race(
+          [
+            new Promise((resolve, reject) => {
+              super
+                .send(method, params)
+                .then((res) => {
+                  this.updateMetrics(true, sendTimestamp, i, method, params);
+                  resolve(res);
+                })
+                .catch((e) => {
+                  const error = parseError(e);
+
+                  reject(error);
+                });
+            }),
+          ].concat(
+            this.stallTimeout
+              ? [
+                  new Promise(async (_, reject) => {
+                    await delay(this.stallTimeout);
+                    reject(new RpcError(RpcError.reasons.Timeout));
+                  }),
+                ]
+              : [],
+          ),
+        );
+      } catch (error) {
+        // If the error thrown is a timeout or non-RPC error, we want to go ahead and throw it.
+        if (
+          (error.type !== ServerError.type ||
+            (error as ServerError).reason === ServerError.reasons.TransactionAlreadyKnown) &&
+          (error.type !== RpcError.type || (error as RpcError).reason === RpcError.reasons.Timeout)
+        ) {
+          throw error;
+        } else {
+          this.updateMetrics(false, sendTimestamp, i, method, params, {
+            type: error.type.toString(),
+            context: error.context,
+          });
+          errors.push(error);
+        }
+      }
+    }
+
+    throw new RpcError(RpcError.reasons.FailedToSend, {
+      provider: this.url,
+      chainId: this.chainId,
+      blockNumber: this.syncedBlockNumber,
+      errors,
+    });
+  }
+
+  private updateMetrics(
+    success: boolean,
+    sendTimestamp: number,
+    iteration: number,
+    method: string,
+    params: any[],
+    error?: { type: string; context: any },
+  ) {
+    const execTime = +((Date.now() - sendTimestamp) / 1000).toFixed(2);
+    this.execTimes.push(execTime);
+    if (success) {
+      this.reliability = Math.min(1, +(this.reliability + SyncProvider.RELIABILITY_STEP).toFixed(2));
+    } else {
+      this.reliability = Math.max(0, +(this.reliability - SyncProvider.RELIABILITY_STEP).toFixed(2));
+    }
+    this.debugLog(
+      success ? "RPC_CALL" : "RPC_ERROR",
+      `#${iteration}`,
+      method,
+      this.cps,
+      execTime,
+      this.reliability,
+      // TODO: Logging params for these methods is for debugging purposes only.
+      ["eth_getBlockByNumber", "eth_getTransactionByHash", "eth_getTransactionReceipt"].includes(method)
+        ? params.length > 0
+          ? params[0]
+          : params
+        : "",
+      error ? error.type : "",
+      error ? error.context : "",
+    );
+  }
+
+  private debugLog(message: string, ...args: any[]) {
+    if (this.debugLogging) {
+      console.log(`[${Date.now()}]`, `(${this.url})`, message, ...args);
     }
   }
 }