+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next Release
 
+- [router] delegate fee calc to chainreader
+- [sdk] integrate chainreader
+- [txservice] introduced provider syncing
+
 ## 0.0.93
 
 - [router] Improved caching and auction response time